+++ Controller/CRUDController.php
@@ -150,7 +150,7 @@ class CRUDController extends Controller
         if ($this->container->has('logger')) {
             return $this->container->get('logger');
         } else {
-            return new NullLogger;
+            return new NullLogger();
         }
     }
 
@@ -289,7 +289,7 @@ class CRUDController extends Controller
                     'sonata_flash_success',
                     $this->admin->trans(
                         'flash_delete_success',
-                        array('%name%' => $this->admin->toString($object)),
+                        array('%name%' => $this->escapeHtml($this->admin->toString($object))),
                         'SonataAdminBundle'
                     )
                 );
@@ -305,7 +305,7 @@ class CRUDController extends Controller
                     'sonata_flash_error',
                     $this->admin->trans(
                         'flash_delete_error',
-                        array('%name%' => $this->admin->toString($object)),
+                        array('%name%' => $this->escapeHtml($this->admin->toString($object))),
                         'SonataAdminBundle'
                     )
                 );
@@ -375,7 +375,7 @@ class CRUDController extends Controller
                         'sonata_flash_success',
                         $this->admin->trans(
                             'flash_edit_success',
-                            array('%name%' => $this->admin->toString($object)),
+                            array('%name%' => $this->escapeHtml($this->admin->toString($object))),
                             'SonataAdminBundle'
                         )
                     );
@@ -397,7 +397,7 @@ class CRUDController extends Controller
                         'sonata_flash_error',
                         $this->admin->trans(
                             'flash_edit_error',
-                            array('%name%' => $this->admin->toString($object)),
+                            array('%name%' => $this->escapeHtml($this->admin->toString($object))),
                             'SonataAdminBundle'
                         )
                     );
@@ -533,7 +533,7 @@ class CRUDController extends Controller
             true;
 
         if ($askConfirmation && $confirmation != 'ok') {
-            $actionLabel = $this->admin->trans($this->admin->getTranslationLabel($action, 'action'));
+            $actionLabel = $batchActions[$action]['label'];
 
             $formView = $datagrid->getForm()->createView();
 
@@ -619,7 +619,7 @@ class CRUDController extends Controller
                         'sonata_flash_success',
                         $this->admin->trans(
                             'flash_create_success',
-                            array('%name%' => $this->admin->toString($object)),
+                            array('%name%' => $this->escapeHtml($this->admin->toString($object))),
                             'SonataAdminBundle'
                         )
                     );
@@ -641,7 +641,7 @@ class CRUDController extends Controller
                         'sonata_flash_error',
                         $this->admin->trans(
                             'flash_create_error',
-                            array('%name%' => $this->admin->toString($object)),
+                            array('%name%' => $this->escapeHtml($this->admin->toString($object))),
                             'SonataAdminBundle'
                         )
                     );
@@ -1088,6 +1088,18 @@ class CRUDController extends Controller
         )) {
             throw new HttpException(400, 'The csrf token is not valid, CSRF attack?');
         }
+    }
+
+    /**
+     * Escape string for html output
+     *
+     * @param string $s
+     *
+     * @return string
+     */
+    protected function escapeHtml($s)
+    {
+        return htmlspecialchars($s, ENT_QUOTES | ENT_SUBSTITUTE, 'UTF-8');
     }
 
     /**